Big rewrite of convective variance transport (CVT)
This adds a new method for variance transport. The two options are
Spectral CVT (SCVT) and Bulk CVT (BCVT). The use of SCVT requires
specifying the number of wavenumber tracers via the crm_ntrc configure
flag. There is also an option to inlcude momentum variance transport in the SCVT mode, but this is currently not implemented for BCVT.

	modified:   components/eam/src/physics/crm/crm_history.F90
	modified:   components/eam/src/physics/crm/crm_input_module.F90
	modified:   components/eam/src/physics/crm/crm_output_module.F90
	modified:   components/eam/src/physics/crm/crm_physics.F90
	modified:   components/eam/src/physics/crm/crmdims.F90
	modified:   components/eam/src/physics/crm/sam/crm_module.F90
	modified:   components/eam/src/physics/crm/sam/crm_variance_transport.F90